rpcmethods: Allow wildcards to be used when subscribing
This provides much more selective subscriptions. You can subscribe to
multiple methods at once and you can also subscribe to the single node
without its children. Or you can subscribe to all nodes with specific
name and so on.

Patterns complicated potentially the work with subscriptions. The
primary issue is that we no longer can check the access rights for the
user because based on the access right database it can be also based on
some sort of wildcard matching and thus check would be really hard or
impossible.

The second issue with wildcard patterns is that pattern needs to be
split to part that applies on tree on the broker and the tree on
sub-broker. This has to be done by taking the wildcard pattern and
finding the longest non-empty path match on the sub-broker's mount
point. Such pattern is the one that should be kept in the broker and
tail should be propagated to the sub-broker. The only special handling
is required for `**`. That is that if tail of the longest match has `**`
you need to use it also as start of the second part of the wildcard, not
the next field after slash.

The question might also be how this affects the rejectNotSubscribed
method. This method drops all subscriptions that can't be propagated to
any user. The question is if by introducing pattern we could break this.
In reality we can't because no matter the rule set the notification is
either propagated to some client or such subscription can't no longer be
required. The proof can be reached by arguing in favor of such
situation. Let's consider that there is such subscription that for some
specific combination of path and method that matches it has no
subscription in upper broker and thus such combination of path and
method would be rejected. Such subscription would not be safe to remove
if there would be some other combination of path and method that matches
it and is propagated by upper broker. To rule that possibility out we
must understand that upper broker subscriptions are handled only up to
the mount point and thus both signals come in eyes of the broker from
the same node (the mount point) and thus subscription on the upper
broker would match both paths the same way. The only difference can be
in method but considering the previously described way distribute
subscriptions to the subbrokers it is clear that method field must be
the same in such case and this it must be the same subscription and thus
such message would not be rejected.
Other way to look at the same proof is that we can consider
subscriptions to include some two groups. Those groups can have overlap.
Let's name the first group A and second group B. The subscription A is
no longer valid because client that added it is no longer present.
Message that is part of A might get rejected because of that, but only
if it is not part of B. The problematic overlap where message is both
part of A and B can't be rejected because such message would be also
part of group B and this propagated. Effectively rejects can happen only
on set A&~B and those by definition do not include subscription B and
this can't cancel it.