rpctransportlayer: remove Datagram
Datagram has issues with reexecution that would require implementation
of some message confirming. The easiest implementation with message
counter pretty much copies TCP design. The advantage would be that we
could do that only some of the messages (for responses because they are
really the issue). The reality is that right now we do not need it and
we probably won't get much of an advantage from it in the current SHV
RPC design. It is just easier to drop it at the moment.