Make signals hidden and directly associated with methods
This changes how we manage signals. They now can't stand alone and
instead must be associated not only with node but also with valid
method.

For compatibility reasons the original field `method` is not considered
to be `ShvName` and thus references not only method name but also signal
name. The new field `Source` was added that signals must use to
associate themself with method.

As a special handling signals names ending with `chng` were set to be
used only by methods that signal same value as they return. For that
reason the original `lschng` signal was renamed to `lsmod` (ls
modification) and associated with `ls` method. The property node's
`chng` signal is now associated with `get` and any signal ending with
`chng` is now allowed.