rpcmethods/broker: Remove rejectNotSubscribed and add subscribe result
The original cleanup method between brokers based on rejectNotSubscribed
is not sufficient because it doesn't cover subscriptions that are never
triggered. The removal of such orphan subscriptions would have to be
done outside of the reject mechanism anyway and that mechanism can also
be automatically triggered when rejectNotSubscribed would be. That
results in to obsolescence of the reject mechanism.