rpcmethods/discovery: do not list signals directly using dir
This removes signals from list provided by `*:dir` method. They are not
only part of their associated method description.

It turns out that signals really only have specific type hint and
otherwise nothing more can be said about them. The result is thus that
we specify them only as mapping of their names to the types they carry.
Nothing else should be required.

This moves also defined signals in the documentation to the sections of
methods they are associated with.

Another significant change here is that `dir` now can be only IMap and
that String parameter results in Bool and not method description.

At the same time the accessGranted was replaced with AccessLevel and
thus string with integer for dir description.