rpcmessage: add Repeat for signals
The reasoning behind adding this meta info to the signal is that
sometimes devices are unable to truly say that some specific signaled
event really happened (such as when enumerating buses) or it might have
been repeat of already reported event. This is for example after devices
restart. The value might have changed in the mean time or it might not.

The alternative considered was to define custom signal name for that but
that would complicate usage of such signals on the client. This way it
is the same signal but it caries info that it might not be trusted as a
trigger but the value it carries is still the correct value.

The name "repeat" was chosen to allow also a future expansion of
broker's subscribe API to allow repeated signals if value doesn't
change. This considered use case wasn't include in the standard right
now.

This value is also propagated to the history. This also includes few
fixed in the history API.