rpcmethods/device: add uptime and reset
These provide info about device uptime and allow systematic request for
device reset. These are features that every device should provide. The
uptime gives a good idea of the device's stability in the network and
reset is operation that should be standardized to prevent it being
implemented on every device a different way.

Both of the methods are designed to be optional in its nature but they
must be implemented.