rpctransportlayer: rename Stream to Block
The Stream name is confusing with concept that this transport layer
works on stream but that also applies to the Serial transport layer.
Thus by renaming them we can talk about both of them as being stream
transport protocols without confusing generic concept of stream
transmission with one of trasport layers name.