rpctransportlayer: try to improve the CAN by addressing collisions
This tries to allow ad hoc connection to the CAN bus without reserving
the address up front. The alternative would be to allow only one ad hoc
connection on specific address but that might be too restrictive. In the
end this is something very close but still allows multiple ad hoc CAN
devices.