rpctransportlayer: split to multiple documents
With increased size of CAN Bus documentation and introduction of
Websockets it no longer makes sense to keep it as one huge page. The
primary page now instead only specifies what we want from transport
layer and then points to separate documents describing how we achieve
that for streams and CAN. This is simply extensible and thus we can
later add any other transport layers if we need them.

And yeah websockets seem to fix what SHV RPC needs exactly (not
surprising when web is the RPC's original home).