Add RPC RI
This was initially introduced in SHVCli while later it got to the
documentation. Now we use it to in general identify specific resources
in text or in SHVCli. The new version of pyshv RPC Broker will also be
using this convention and thus it is makes sense to define it here.

The alternative would be to maintain it in pyshv documentation but
it makes more sense to place it here, because these resource identifiers
are used in this documentation as well and thus should not be without an
explanation.