rpcmethods: Remove signatures
Signatures are no longer releavant. The presence of parameter and result
in dir description is signaled by next table that provides info about
parameters and result. The presence in the first table is duplicate
information that is no longer relevant.

This doesn't cover issue that for some methods it is not clear what
dir's param and result should contain but that must be filled in once we
define the type descriptions in the future.