rpcmethods/discovery: allow ls and dir to be extended in the future
The check for the method and child node existence with ls and dir can be
in the future extended this way to provide some additional info instead
of just `true`. In case of dir this even increases compatibility with
previous implementations.

The idea for the future is to allow aliasing of node and method names
with IDs. The ls and dir existence query in such case would return
integer ID instead of `true`.