Update concept, tweak language and replace Rpc with RPC
This is cleanup of the documentation to make sure that it is consistent.
The concept article is updated to cover latest tweaks to the standard.
The Rpc is being replaced with RPC because that is shortcut and the
original Rpc* names were from the code which is irelevant to the
standard and the correct one is RPC.