rpcmethods/file: Add MaxSize to stat method
On platforms with limited memory pool it is hard to implement write
operations because in general in SHV it is not clear if message is valid
until it is fully received. For that reason implementations must buffer
all received data and thus there must be limit on maximum size that can
be handled. The caller can't know this limit and thus it would have to
be ether tied to the PageSize or it would have to be somehow discovered.
This way it is directly discloused.

Notice that this limit is not in any way tied to the PageSize and thus
it is absolutely possible that MaxSize is smaller than PageSize as well
as other way around.