rpcmethods/discovery: mention that dir result should not change
The issue with added and removed methods is that there is no standard
way that could be signaled. It can also introduce race condition where
method is discovered and removed before call. The other way around where
method is added later down the line is problematic as well because tools
using SHV would have to poll dir to detect this.