rpcmethods/history: update old way to filter signals
The commit bedaca9643441c7ab28777d45bd726eadb35105e changed the Broker's
API to use RPC RI instead of dedicated fields. This made getLog out of
sync in the design as well as in terms of links. This correctly uses
same RPC RI as subscription but at the same time provides hint that it
might not be a good idea to use it to filter based on the path.