rpcmessage: obsole some and change how UserCode is documented
The obsolete errors are not used at the moment and from point of RPC
they makes no sense.

Some of the implementation specific errors are switched in here to be
documented as the error reserved to the implementation. They won't be
used in the communication.

The user specific error codes are modified to be application specific
and all of them are now tied to be more specific of MethodCallException.