rpcmethods/history: Change the .records:span fields meaning
This is required because otherwise corner case of no records is not well
defined. By always signaling ID right after the latest record ensures
that we can signal no record with `[0,0,x]`. The alternative would be to
consider `0` ID as a hole but implementations would have to make sure to
never use this ID because upper histories could fetch empty list of
records with record 0 as no record and would not ask for it again in the
future and thus this record 0 would be ignored.