rpctypes: Add type hints definition
The type hints definition should be used to hint on the expected
parameter and result of the methods. It is designed to be parsed by
decision tree.

This commit also defines parameter and result type for all common RPC
methods definitions. It is included in the single table as a reference
for the 'dir' description of that method. The signals are also described
this way and thus their description united across the documentation.

The 'dir' method's result now directly references type descriptions as
the format to be used.