rpcmethods/broker: convert client info from Map to IMap
This is incompatible change!

We do this because this way the API is consistent with others (such as
`*:dir`).

The effect should be minimal because this API is intended to be used by
broker administrators and not by applications. It might be used in some
test scenarios and thus those will be affected but this change might
benefit them more than not doing this change.