rpctypes: Add type hints definition (#51)
* rpctypes: Add type hints definition

The type hints definition should be used to hint on the expected
parameter and result of the methods. It is designed to be parsed by
decision tree.

This commit also defines parameter and result type for all common RPC
methods definitions. It is included in the single table as a reference
for the 'dir' description of that method. The signals are also described
this way and thus their description united across the documentation.

The 'dir' method's result now directly references type descriptions as
the format to be used.

* rpctypes: add grammar

This makes the format more concrete and exact. It can also be used as a
reference for the parser implementation.

* github/workflow: remove fix-pagetoc branch condition

This was introduced in 9dba97e65acfe842fbc1b3474f364d84083f9e15 and is
clearly a left over from debugging.

* rpcmethods/broker: fix subscriptions listing examples

The old list based way of providing subscriptions was left in the
examples when we switched the format to the Map. This fixes these
examples.

This also fixes the documentation for the subscription field of
.broker/currentClient:info.

* rpcmethods/broker: convert client info from Map to IMap

This is incompatible change!

We do this because this way the API is consistent with others (such as
`*:dir`).

The effect should be minimal because this API is intended to be used by
broker administrators and not by applications. It might be used in some
test scenarios and thus those will be affected but this change might
benefit them more than not doing this change.