rpcmethods/discovery: add isUpdatable flag
The issue we are trying to solve here is that sometimes it is desirable
that all items in compound types are provided while in some other cases
it is more desirable to provide smaller message with only updated items.
This difference might not be easily addressed by type hint because we
commonly required that type of the `:get` result is the same as `:set`
parameter. But we want to allow these smaller update messages.