rpcmessage: add delayed response (#60)
* rpcmessage: add delayed response

It is common that some method calls can take a considerably longer than
the message transfer. Our approach so far was, and still is, to send
request again after some time as we expect that either request or
response was lost. This timeout is crucial because if it is too short
then we could unintentionally instruct device to perform call multiple
times which is more likely for calls running longer and thus the effect
would be even larger. This is positive feedback loop that can flood the
call executing device.

This provides a way to side step this issue. The message about delayed
response should be sent periodically to inform caller that neither
request nor response got lost and thus that there is no need to send
another request. The period should be chosen such that multiple delayed
response messages are sent in the window of expected call timeout so the
likelihood of delivering information about delayed response is
increased. Caller should start using Abort Request with false after it
receives delayed request and thus after initial period of more intensive
delayed response sending or right after first abort request the period
can be much longer.

This introduction also requires two additional errors. The
RequestInvalid is to signal that that either call was aborted after
Abort with `true` or that there is no such request at the moment. The
second error is the expectation that devices can have only limited
number of slots for the parallel requests and thus they need some way to
inform caller that they are just under pressure and that another
attempt later might be successful.

* rpcmessage: remove Part message attribute

It was never used and is no longer planned to be used.