rpctypes: remove `|b` from !dir
The `|b` is included in the description of the method and it is better
to have it defined that way to make the different calling types explicit
and not hiden in this type definition.

By removing this the double definition of boolean is removed and only
the definition in the `dir` result is used.