rpctypes: remove `|b` from !dir (#63)
* rpctypes: remove `|b` from !dir

The `|b` is included in the description of the method and it is better
to have it defined that way to make the different calling types explicit
and not hiden in this type definition.

By removing this the double definition of boolean is removed and only
the definition in the `dir` result is used.

* shvrpcconcepts: improvide the delayed response description

This should make more clear who is sending which message and what is the
expected behavior.

* workflows: update used mdbook version to 0.4.49

Just a maintenance update to use latest version.

* mdbook: add linkcheck to ensure that links are correct

This will discover broker links in CI.

* rpcmessage: CallerIds and RevCallerIds allow allone Int as well

This is what we have in all implementations to safe few bytes in message
but it wasn't correctly covered in the SHV standard.

* rpctypes: fix invalid standard getLogP and getLogR

Just two minor issues that cause type to not match what is defined in
the documentation.