.github: ensure deploy is performed only on push to master
Without this if deploy is performed even for merge requests, thus
overwriting the docs with every merge request. On the other hand,
it might be useful to do the build job on merge request, so just
add if statement to deploy instead of disabling the entire workflow
for merge requests.

Signed-off-by: Michal Lenc <michallenc@seznam.cz>