rpcmethods/history: getLog result Ref no longer provides defaults
The option to allow usage of any closest matching record for defaults is
not ideal as it is unclear which reference should be used. Should we use
the closest one that matches at least one of the fields? Or should we
use the closes one matching all fields? The client thus has to keep all
combinations and not just unique combination references as device migth
reference any previous record. The effect is thus more complex client
for chance that device might be even more complext (which is unlikely).

This thus changes the requirement for setting default from reference to
setting the value of those items. Client thus should ignore Signal,
Source, and Path if Ref is used.