rpcmethods/history: modify the records fetch method
This introduces two changes tied to each other.

The ID is added to the records. This is required because otherwise
storage that contains holes can't be replicated in the fetcher. Consider
storage where every tenth ID is valid (the rest are holes). The caller
with requires `[0,100]` thus will get only ten results. Without
knowledge of underlying holes it can't decide on the exact valid ID. The
holes in the records are essential feature to allow low level mapping
of offsets to the IDs without mapping table.

The ref is added to the records that mimics the Ref in getLog result.
This also allows less than requested records to be returned from the
device. The issue this solves is that device might not be able to buffer
and send all the requested records and there is no way to know the limit
upfront as it depends on the queried records. With this change the fetch
can return less records than requested and caller should ask for more
by using the ID of the last records.