src/rpcmethods/file.md: describe file node for MTD layer (#69)
* src/rpcmethods/file.md: add MaxRead parameter to stat method

The field imposes the limitation on read operation. It defines the
maximum number of bytes read can return. This allows the device
to avoid large buffering.

Signed-off-by: Michal Lenc <michallenc@seznam.cz>

* src/rpcmethods/file.md: describe file node for MTD layer

This ensures the file node is ready to handle the behavior of
direct access to MTD layer. It basically requires the writes to be
in a sequential order within one erase page, otherwise the data
may be corrupted.

Signed-off-by: Michal Lenc <michallenc@seznam.cz>

---------

Signed-off-by: Michal Lenc <michallenc@seznam.cz>