rpcri: Remove last paragraph
This is out of place statement. It has nothing to do directly with SHV
RI but rather with access control configuration of the broker. The
closest definition for this is broker API but that doesn't require any
specific access assignment. Actually it is up to the broker
implementation how access levels are assigned.