rpcmessage: rename Params to Param and RPC message needs meta-data
This just clarifies few definitions. The message caries actually one
complex parameter and defining them as Params is incosistent with rest
of the documentation.

This has mostly no real impact as it is only a convention tweak. The
existing implementations are encouraged to keep Params and switch only
with some major release to introduce API breakage. The difference is
negligible.

The second fix here is that the sentence saying that meta-data can be
specified is wrong as they have to be provided, just some of the are
optional.